+++ src/Oro/Bundle/ApiBundle/Util/DependencyInjectionUtil.php
@@ -12,9 +12,12 @@ use Symfony\Component\DependencyInjection\Reference;
  */
 class DependencyInjectionUtil
 {
+    /** the name of DIC tag for Data API processors */
+    public const PROCESSOR_TAG = 'oro.api.processor';
+
     /**
      * @internal never use this constant outside of ApiBundle,
-     *           to recive and update the configuration use getConfig and setConfig methods.
+     *           to receive and update the configuration use getConfig and setConfig methods.
      */
     public const API_BUNDLE_CONFIG_PARAMETER_NAME = 'oro_api.bundle_config';
 
@@ -86,7 +89,7 @@ class DependencyInjectionUtil
      *
      * @return mixed
      *
-     * @throws LogicException is the reuested attribute does not exist in $attributes array
+     * @throws LogicException is the requested attribute does not exist in $attributes array
      */
     public static function getRequiredAttribute(array $attributes, $attributeName, $serviceId, $tagName)
     {
@@ -182,8 +185,8 @@ class DependencyInjectionUtil
         string $requestType
     ) {
         $processorDef = $container->getDefinition($processorServiceId);
-        $tags = $processorDef->getTag('oro.api.processor');
-        $processorDef->clearTag('oro.api.processor');
+        $tags = $processorDef->getTag(self::PROCESSOR_TAG);
+        $processorDef->clearTag(self::PROCESSOR_TAG);
 
         foreach ($tags as $tag) {
             if (empty($tag['requestType'])) {
@@ -191,7 +194,7 @@ class DependencyInjectionUtil
             } else {
                 $tag['requestType'] .= '&!' . $requestType;
             }
-            $processorDef->addTag('oro.api.processor', $tag);
+            $processorDef->addTag(self::PROCESSOR_TAG, $tag);
         }
     }
 }